% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColSnake.R
\name{ColSnake}
\alias{ColSnake}
\title{rename Data Frame Columns to Snake Case}
\usage{
ColSnake(df)
}
\arguments{
\item{df}{a data frame whose column names need to be converted to snake_case.}
}
\value{
a data frame with renamed columns in snake_case.
}
\description{
the function renames all columns in a data frame to snake_case.
}
\examples{
 
df <- data.frame("Column One" = 1:3, "Column.Two" = 4:6)
ColSnake(df)
# Returns a data frame with columns "column_one" and "column_two".
}
