% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TooSnake.R
\name{TooSnake}
\alias{TooSnake}
\title{Convert a String to Snake Case}
\usage{
TooSnake(string)
}
\arguments{
\item{string}{A string to be converted to snake_case.}
}
\value{
A string in snake_case format.
}
\description{
This function changes a string to snake_case by:
\itemize{
\item changing characters to lowercase
\item spaces become underscrores
}
}
\examples{
TooSnake("a /CoLuMn /NaMe")
# returns: "a_column_name"
}
