% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColSnake.R
\name{ColSnake}
\alias{ColSnake}
\title{rename Data Frame Columns to Snake Case if using a tibble convert to dataframe using base::as.data.frame()}
\usage{
ColSnake(df)
}
\arguments{
\item{df}{a data frame whose column names need to be converted to snake_case.}
}
\value{
a data frame with renamed columns in snake_case.
}
\description{
the function renames all columns in a data frame to snake_case.
}
